pptx-backend/requirements.txt
pandas
numpy
python-pptx
openpyxl
lxml
Pillow
requests
pip-system-certs
fastapi
uvicorn